+++ angular/directives/data_listing/data_listing.js
@@ -1,0 +1,17 @@
+(function(){
+	"use strict";
+
+	angular.module('app.directives').directive( 'DataListing', function() {
+
+		return {
+			restrict: 'EA',
+			templateUrl: 'views/directives/data_listing/data_listing.html',
+			controller: 'DataListingCtrl',
+			link: function( $scope, element, $attrs ){
+				//
+			}
+		};
+
+	});
+
+})();