+++ app/Console/Kernel.php
@@ -16,6 +16,7 @@ class Kernel extends ConsoleKernel
         'App\Console\Commands\Inspire',
         'App\Console\Commands\AngularFeature',
         'App\Console\Commands\AngularDialog',
+        'App\Console\Commands\AngularDirective',
         'App\Console\Commands\AngularService',
         'App\Console\Commands\AngularFilter',
         'App\Console\Commands\AngularConfig',